pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "printing-service"
version = "0.2.0"
description = "A Python library providing two brokers for interacting with CUPS servers: a subprocess-based CLI broker and a high-level pycups broker."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
    { name="Mathias", email="mathiasjohnbeck@gmail.com" }
]

dependencies = [
    "pycups>=2.0.4"
]

[project.urls]
Homepage = "https://github.com/Rigshospitalet-KFNM/printing_service"

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
